+++ src/app/(auth)/signup/layout.tsx
@@ -1,23 +1,20 @@
 'use client';
 
-import BackBtn from '@/app/components/common/BackHeader';
-import NextBtn from '@/app/components/common/NextBtn';
+import BackBtn from '@/app/components/common/ui/BackHeader';
+import NextBtn from '@/app/components/common/ui/NextBtn';
 import { useSignUpStore } from '@/store/SignupStore';
 
-
 export default function Layout({ children }: { children: React.ReactNode }) {
-  
-  const step = useSignUpStore((state)=> state.step)
-  const setStep = useSignUpStore((state) => state.setStep)
+  const step = useSignUpStore((state) => state.step);
+  const setStep = useSignUpStore((state) => state.setStep);
 
-  const isNextEnabled = useSignUpStore((state) => state.isNextEnabled)
+  const isNextEnabled = useSignUpStore((state) => state.isNextEnabled);
 
   const goNext = () => {
-    if(step<6){
-      setStep (step + 1)
+    if (step < 6) {
+      setStep(step + 1);
     }
-  }
-  
+  };
 
   // 경로에 따른 제목 설정
   const changeTitle = () => {
@@ -29,13 +26,18 @@ export default function Layout({ children }: { children: React.ReactNode }) {
     if (step === 6) return '관심 카테고리 설정';
     return '회원가입';
   };
-  
+
   return (
     <>
-      <div className="flex flex-col min-h-screen">
+      <div className="flex min-h-screen flex-col">
         <BackBtn title={changeTitle()} />
         <div>{children}</div>
-        <NextBtn label='다음' onClick={goNext} disabled={!isNextEnabled} className={`${isNextEnabled ? 'bg-[#222222]' : 'bg-[#c4c4c4]'}`}/>
+        <NextBtn
+          label="다음"
+          onClick={goNext}
+          disabled={!isNextEnabled}
+          className={`${isNextEnabled ? 'bg-[#222222]' : 'bg-[#c4c4c4]'}`}
+        />
       </div>
     </>
   );
