+++ src/app/components/signup/UserInfo.tsx
@@ -5,8 +5,8 @@ import checkGray from '/public/checkGray.svg';
 import showBtn from '/public/showBtn.svg';
 import { useEffect, useState } from 'react';
 import { useSignUpStore } from '@/store/SignupStore';
-import Input from '../common/Input';
-import Button from '../common/Button';
+import Input from '../common/ui/Input';
+import Button from '../common/ui/Button';
 
 export default function UserInfo() {
   const name = useSignUpStore((state) => state.name);
